+++ src/gutenberg2zim/__main__.py
@@ -1,4 +1,5 @@
-from gutenberg2zim.entrypoint import main
-
 if __name__ == "__main__":
-    main()
+    print(  # noqa: T201
+        "This project has moved to https://github.com/openzim/papers on Github, "
+        "papers2zim on Pypi"
+    )