+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 as of 2.0.0.
 
+## Unreleased
+
+### Changed
+
+- Renamed to papers2zim, this gutenberg2zim package is now obsolete
+
 ## [3.0.1] - 2025-11-24
 
 ### Added